stats: explicitly cast "time_t" to "long"
Fixes warning on mingw32.

+++ src/stats.c
@@ -116,6 +116,6 @@ int stats_print(struct client *client)
 		      (long)g_timer_elapsed(stats.timer, NULL),
 		      (long)(pc_get_total_play_time() + 0.5),
 		      stats.song_duration,
-		      db_get_mtime());
+		      (long)db_get_mtime());
 	return 0;
 }
